plant-disease-detection/requirements.txt
streamlit
torch
torchvision
numpy
Pillow
